fix(layout): harden web_lift diagnostic to a safe-aligned byte read
The `web_lift` text-error diagnostic read the error enum's first u32 via
`*(&e as *const _ as *const u32)` — UB on a 1-aligned or <4-byte enum (same
class as the screencap POD bug). The marker only needs the discriminant, so
read a single `*const u8` (always aligned + in-bounds).

Audit note: reviewed transmute / from_raw_parts / raw scalar derefs across
core, css and layout. The rest is required unsafe — the gl.rs / json.rs ptr+len
wrappers are repr(C) C-ABI Vec/String types, transform.rs is SIMD intrinsics,
host_invoker.rs / callbacks.rs are runtime fn-ptr transmutes, css/macros.rs is
a Vec move-out. `parse_format_pod` (screencap) was the only genuinely
convertible one and is now safe slice access.

+++ layout/src/solver3/fc.rs
@@ -2743,7 +2743,10 @@ fn layout_ifc<T: ParsedFontTrait>(
             #[cfg(feature = "web_lift")]
             unsafe {
                 crate::az_mark((0x60688) as u32, (0x000000EEu32) as u32);
-                crate::az_mark((0x6068C) as u32, (*(&e as *const _ as *const u32)) as u32);
+                // Read the error's first byte (discriminant) for the marker — a
+                // `*const u8` read is always aligned + in-bounds; the old
+                // `*const u32` read was UB on a 1-aligned / <4-byte enum.
+                crate::az_mark((0x6068C) as u32, (*(&e as *const _ as *const u8)) as u32);
             }
             // Font errors should not stop layout of other elements.
             // Log the error and return a zero-sized layout.
